fix(ingestion): issue #131 P5 split source builders
