+++ docs/architecture.md
@@ -86,7 +86,7 @@ and respect the handoff invariants below.
 
 **Handoff invariants** (must hold for every new fetcher / storage table):
 
-1. **Bronze must be writable before silver in any new fetcher.** A scraper that lands silver without first writing the raw HTTP body to a `*_raw` table is a regression — there's no recovery path when the parser bug shows up months later. Use the existing `*_with_raw` companion pattern: scraper returns `(parsed_obs, raw_payload, request_params)`, fetcher populates `RawSeries.raw_payload + content_hash + request_params_json`, orchestrator's `_capture_obs_raw` side-writes the audit row before normalize. Issue #131 closes the legacy time-series follow-up routes; remaining work is physical cleanup of the source registry surface.
+1. **Bronze must be writable before silver in any new fetcher.** A scraper that lands silver without first writing the raw HTTP body to a `*_raw` table is a regression — there's no recovery path when the parser bug shows up months later. Use the existing `*_with_raw` companion pattern: scraper returns `(parsed_obs, raw_payload, request_params)`, fetcher populates `RawSeries.raw_payload + content_hash + request_params_json`, orchestrator's `_capture_obs_raw` side-writes the audit row before normalize. Issue #131 closed the legacy time-series follow-up routes and split the source builder surface by geography.
 2. **Silver columns cannot reference data not derivable from bronze.** If silver carries a column, the parser must be able to compute it from the stored raw payload alone — otherwise replay produces a different projection than live ingest and the audit lane is a lie. Synthetic columns sourced from a separate seed table (`obs_family_id` joins via `concept_map`, `subject_aliases` lookups) are fine because they're context-derivable, not payload-derivable; provider metadata that exists only in the live response (e.g. response headers, ETags, server clock) is not. The P3 round-trip tests (`tests/test_bronze_replay.py`) enforce this for time-series sources.
 3. **Gold must be deterministically derivable from silver.** No gold view may aggregate or rank against bronze directly — bronze is for audit + replay, not for queries. The cross-source priority + PIT projection logic stays in silver-side code (`resolve_indicator` reads `indicator_vintages` + `concept_map`; `list_items` reads `subjects` + projection tables). This keeps the audit lane decoupled from the query path so re-canonicalizing bronze (e.g. dropping a newly-discovered envelope echo) doesn't risk breaking gold reads.
 
@@ -114,7 +114,7 @@ and respect the handoff invariants below.
 - **Resolution:** `resolve_indicator(concept_id, date=None, as_of=None)` ranks sources by `concept_map.priority`, returns primary + alternates. The `as_of=YYYY-MM-DD` parameter (issue #114 P3) selects the latest vintage with `vintage_date <= as_of` — true point-in-time reads. `resolve_indicator_history(as_of=...)` reconstructs the projection at the requested cutoff. Calendar fallback (#114 P3): when the latest projection has fewer than 24 rows, results are augmented with `cal_econ_event.actual` rows for that concept's release_family, tagged `provenance='calendar'` (downstream filterable). Bridge lives in `_CALENDAR_FALLBACK_KEYWORDS` in `storage/queries/indicator.py` — concept_id → (event title keywords, country_code).
 - **Status:** architecture complete; first-pass ingestion bootstrap underway. DB starts empty; Phase 1 (top US macro, 10 indicators) is the immediate next milestone, not more code.
 - **Audit lane** (issue #69 slice 1, completed by issue #116): `obs_raw` mirrors `cal_econ_raw` — one row per HTTP response per `(source, series_id)`, idempotent on `(source, series_id, content_hash)`. Per-source canonicalization (`ingestion/timeseries/canonicalize.py`) hashes only the sorted observations, dropping query-time envelope echoes (`observation_start` / `realtime_*` / `responseTime` / SDMX `header.prepared`) so a daily refresh with a sliding `start_date` dedupes through INSERT OR IGNORE. The orchestrator's `_capture_obs_raw` side-writes the audit row before normalize whenever the fetcher populates `raw_payload + content_hash`. Re-projection helpers (`_parse_fred_observations` / `_parse_X` in BLS+SDMX+MOF+AISI+ISM+Redbook+sentix) mirror the calendar `cal_econ_raw` replay path — fix a parser, replay raw, zero quota cost.
-  - **Bronze coverage status (issue #131 in progress):** *Wired+running* on the routine fetcher pipeline (`_build_fetcher_source` / `_build_vintage_fetcher_source` + `*_with_raw` companion + canonicalize fn): fred_daily / fred_nondaily / fred_full, fred_vintages, imf_vintages, bls / imf / eia (including cache-hit + FRED fallback), ecb / bundesbank / mof_jp / aisi / ism / redbook / sentix, nyfed_rates, treasury_fiscal, eurostat (JSON-stat path), oecd, worldbank, worldbank_catalog. *Remaining issue #131 cleanup:* retire the generic `execute()` branch and physically split `sources.py` by geography. *Deferred*: bis (CSV body, JSON-shape canonicalization not applicable). The one-shot `scripts/backfill_concept.py` historical sweep also bypasses bronze and can't be replayed (no kept payload); going forward, routine refresh accumulates the audit row.
+  - **Bronze coverage status (issue #131 complete):** *Wired+running* on the routine fetcher pipeline (`_build_fetcher_source` / `_build_vintage_fetcher_source` + `*_with_raw` companion + canonicalize fn): fred_daily / fred_nondaily / fred_full, fred_vintages, imf_vintages, bls / imf / eia (including cache-hit + FRED fallback), ecb / bundesbank / mof_jp / aisi / ism / redbook / sentix, nyfed_rates, treasury_fiscal, eurostat (JSON-stat path), oecd, worldbank, worldbank_catalog. Source builder ownership lives in `ingestion/sources/{us,eu,jp,cn,global}.py`; the orchestrator uses the fetch / normalize / validate / deduplicate / store pipeline for every registered source. *Deferred*: bis (CSV body, JSON-shape canonicalization not applicable). The one-shot `scripts/backfill_concept.py` historical sweep also bypasses bronze and can't be replayed (no kept payload); going forward, routine refresh accumulates the audit row.
 
 ### 2. Documents (government reports + central-bank comms)
