+++ src/ingestion/sources/cn.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+from typing import Any
+
+
+SOURCE_METHODS: dict[str, Any] = {}
